+++ lib/php/functions.php
@@ -476,5 +476,35 @@ function sanitize($string = '', $is_filename = false) {
 
 }
 
+/**
+ * @brief Slugify a string for url usage
+ * @param string $text string to slugify
+ * @return slugified string
+ */
+function slugify($text) {
+
+  // replace non letter or digits by -
+  $text = preg_replace('~[^\\pL\d]+~u', '-', $text);
+
+  // trim
+  $text = trim($text, '-');
+
+  // transliterate
+  $text = iconv('utf-8', 'us-ascii//TRANSLIT', $text);
+
+  // lowercase
+  $text = strtolower($text);
+
+  // remove unwanted characters
+  $text = preg_replace('~[^-\w]+~', '', $text);
+
+  if (empty($text)) {
+    return 'n-a';
+  }
+
+  return $text;
+
+}
+
 
 ?>